refactor: rename files to put all curves in one place
