+++ src/pynurbs/curves/curves.py
