+++ tests/curves/test_spline.py
@@ -3,8 +3,8 @@ from copy import copy
 import numpy as np
 import pytest
 
+from pynurbs.curves.curves import Curve
 from pynurbs.knotspace import GeneratorKnotVector, KnotVector
-from pynurbs.responsive.curves import Curve
 
 
 @pytest.mark.order(34)
@@ -12,7 +12,7 @@ from pynurbs.responsive.curves import Curve
     depends=[
         "tests/test_knotspace.py::test_end",
         "tests/test_basis_functions.py::test_end",
-        "tests/responsive/test_beziercurve.py::test_end",
+        "tests/curves/test_bezier.py::test_end",
     ],
     scope="session",
 )
