+++ tests/operations/test_customstruc.py
@@ -114,9 +114,9 @@ class CustomPoint:
     depends=[
         "tests/test_knotspace.py::test_end",
         "tests/test_basis_functions.py::test_end",
-        "tests/responsive/test_beziercurve.py::test_end",
-        "tests/responsive/test_splinecurve.py::test_end",
-        "tests/responsive/test_rationalcurve.py::test_end",
+        "tests/curves/test_bezier.py::test_end",
+        "tests/curves/test_spline.py::test_end",
+        "tests/curves/test_rational.py::test_end",
     ],
     scope="session",
 )
